Inclusão da Extensão CEP e reestruturação do projeto
