+++ src/W4G.Extensions/Interfaces/ICepService.cs
@@ -1,0 +1,10 @@
+using W4G.Extensions.Models;
+
+namespace W4G.Extensions.Interfaces
+{
+    internal interface ICepService
+    {
+        bool Validate(string cep);
+        Endereco Info(string cep);
+    }
+}
