+++ src/W4G.Extensions/Models/Endereco.cs
@@ -1,0 +1,14 @@
+namespace W4G.Extensions.Models;
+public class Endereco
+{
+    public string Cep { get; set; }
+    public string Logradouro { get; set; }
+    public string Complemento { get; set; }
+    public string Bairro { get; set; }
+    public string Localidade { get; set; }
+    public string Uf { get; set; }
+    public string Ibge { get; set; }
+    public string Gia { get; set; }
+    public string Ddd { get; set; }
+    public string Siafi { get; set; }
+}
