+++ tests/W4G.Extensions.Test/Extensions/Enums/EnumTeste.cs
