+++ tests/W4G.Extensions.Test/Extensions/Strings/NoAccentsTest.cs
@@ -1,0 +1,31 @@
+using W4G.Extensions.Strings;
+
+namespace W4G.Test
+{
+    [TestClass]
+    public class NoAccentsTest
+    {
+        private void NoAccents(string expected, string atual) => Assert.AreEqual(expected, atual.NoAccents());
+
+        [TestMethod]
+        public void NoAccents()
+        {
+            NoAccents("aeiouc$&aAj", "áéíóúç$&aAj");
+            NoAccents("Joao", "João");
+            NoAccents("Acucar", "Açucar");
+        }
+
+        [TestMethod]
+        public void NoAccentsEmpty()
+        {
+            NoAccents("", "");
+        }
+
+        [TestMethod]
+        public void NoAccentsNull()
+        {
+            NoAccents(null, null);
+        }
+
+    }
+}
